Pass directory info when loading maps
Pass along the directory when loading a map.  This is a breaking change
for anything that calls LoadMapRaw() directly.

Tileset filenames in the source map files are referenced with relative
paths, not absolute paths.  In order to properly load maps in
directories other than the current working directory, directory
information needs to be passed along to the tileset loading code.